requirements.txt
flask
pypdf
pdfplumber
pdfminer.six
pymupdf
Pillow
pytesseract
easyocr
boto3
python-dotenv
llama-parse
docling
anthropic
openai
google-cloud-aiplatform
